update telegram bot version

+++ src/Commands/HelpCommand.cs
@@ -18,7 +18,7 @@ namespace TelegramBotTemplate.Commands
             _serviceProvider = serviceProvider;
         }
 
-        public async Task Execute(IChatService chatService, long chatId, int userId, int messageId, string? commandText)
+        public async Task Execute(IChatService chatService, long chatId, long userId, int messageId, string? commandText)
         {
             await chatService.SendMessage(chatId, "TODO: Create a todo command");
         }
